add goods test

+++ springcloud/goodserver/src/main/java/Goods/ServiceImpl/GoodsServiceImpl.java
@@ -6,6 +6,8 @@ import Goods.Service.GoodsService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.util.Objects;
+
 @Service
 public class GoodsServiceImpl implements GoodsService {
 
@@ -17,8 +19,9 @@ public class GoodsServiceImpl implements GoodsService {
         return  goodsDao.findone(id);
     }
     @Override
-    public void saveGoods(Goods goods) {
-        goodsDao.Save(goods);
+    public Integer saveGoods(Goods goods) {
+        if(Objects.equals(goods.getTitle(),"") || Objects.equals(goods.getUsername(), "")){ return -1; }
+        return goodsDao.Save(goods);
     }
 
     @Override
@@ -27,6 +30,9 @@ public class GoodsServiceImpl implements GoodsService {
     }
     @Override
     public String deleteGoodsById(int Good_id){
+        if(Good_id == -1){
+            return "不存在该商品！";
+        }
         return goodsDao.deleteGoodsById(Good_id);
     }
 }
